+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - 全面规范化了代码，以符合 [NoneBot](https://github.com/nonebot/nonebot2/issues/3339) 发布标准。
 - 数据存储迁移至 NoneBot 官方 `localstore` 插件。
+- 修复了在使用 Console Adapter 时对话冷却的错误。
 
 ### v1.1.4