+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # 更新日志
 
-### v1.1.5 (latest)
+### v1.1.5
 
 - 全面规范化了代码，以符合 [NoneBot](https://github.com/nonebot/nonebot2/issues/3339) 发布标准。
 - 数据存储迁移至 NoneBot 官方 `localstore` 插件。
 - 修复了在使用 Console Adapter 时对话冷却的错误。
+- 修复了刷新对话后，仍然提示对话超时的错误。
 
 ### v1.1.4